% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BisectKmeansClustering.r
\name{BisectKmeansClustering}
\alias{BisectKmeansClustering}
\title{Bisect k-means clustering of the data}
\usage{
BisectKmeansClustering(Inputdata, cluster.size = 0.5)
}
\arguments{
\item{cluster.size}{The fraction of pixels in individual cluster. Particular cluster will not split further if 
contains this fracion of data-points with respect total data-points.}

\item{x}{A numeric matrix of data.}
}
\value{
A vector of integers indicating the cluster to which each point is allocated.
}
\description{
Perform bisect k-means clustering on a data-matrix.
}
\details{
The data given by x is clustered by the k-means algorimthm based correlation distance. The 
number of clusters automatically decides according calinhara internal cluster validity index.
The split of cluster into another cluster stop if the size of become less than cluster.size vlaue which the
fraction of data points into compare compare to total size of input data.
}
\examples{
x <- rbind(matrix(rnorm(500,sd=0.3),ncol=2),
          matrix(rnorm(500,mean=1,sd=0.3),ncol=2))
cl <- BisectKmeansClustering(x)
plot(x,col=cl)               

}
\seealso{
optimalClust
}
