% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantitativeLISAmap.r
\name{ClusteredLISAimgQA}
\alias{ClusteredLISAimgQA}
\title{Quantative analysis of segemented image and LISA map}
\usage{
ClusteredLISAimgQA(ClusteredImg, LISAmap, plot = "T")
}
\arguments{
\item{ClusteredImg}{two-dimensinal clustered image}

\item{LISAmap}{two-dimensional LISA map}
}
\value{
fraction of pixels from different clusters fall into different zones of the LISA map
}
\description{
The function calculates the  fraction of pixels from different clusters fall into different zones of the LISA map
}
