% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLISAmap.r
\name{CreateLISAmap}
\alias{CreateLISAmap}
\title{Creates LISA map of input image}
\usage{
CreateLISAmap(Ionimage, distthrs = 5, plot = "T", mask = NULL)
}
\arguments{
\item{Ionimage}{two-dimensinal data matrix to create LISA map}

\item{distthrs}{lag distance value}

\item{plot}{return LISA image in output}

\item{mask}{mask image of tissue to exclude backaground region in the calculation}
}
\value{
two-dimensional LISA map
}
\description{
Creates LISA image of data matrix
}
