% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpatialCorrelogram.r
\name{CreateSpatialCorrelogram}
\alias{CreateSpatialCorrelogram}
\title{Spatial correlogram for Moran's I}
\usage{
CreateSpatialCorrelogram(w, x, distr = 2, Fig = "T")
}
\arguments{
\item{w}{a two-dimensional matrix with xy-coordinates of input variable}

\item{x}{a numeric vector}

\item{distr}{maximum lag order}

\item{plot}{logical operator to decide make spatial correlogram plot}
}
\value{
two-dimension matrix of lag distance and corresponding autocorrelation value
}
\description{
creates spatial correlogram of input variable at different lag distances
}
