% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianFilterR.r
\name{medianFilterR}
\alias{medianFilterR}
\title{Median filtering of image}
\usage{
medianFilterR(sampleMat)
}
\arguments{
\item{sampleMat}{image for median filtering}
}
\value{
returns median filtered image
}
\description{
performs median filtering of the image with window size 3 x 3
}
