% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BisectKmeansClustering.r
\name{optimalClust}
\alias{optimalClust}
\title{Find optimal number of clusters}
\usage{
optimalClust(Inputdata)
}
\arguments{
\item{Inputdata}{data to calculate number of clusters}
}
\value{
number of clusters
}
\description{
function returns optimal number of clusters based on calinhara internal cluster validity index method for k-means correlation distance based clustering method
}
