% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProprocessingMSIdata.r
\name{PreprocessingMSIData}
\alias{PreprocessingMSIData}
\title{Preprocessing of MSI data}
\usage{
PreprocessingMSIData(folderpath, removePeaks = 0.2, edgecorrection = T)
}
\arguments{
\item{folderpath}{path of MSI data folder. At the moment our function only supports the MSI data from in Analyze 7.5 format.}

\item{removePeaks}{remove the bins from MSI data with lower coverage inside tissue below defined threshold}

\item{edgecorrection}{remove high intensity pixels from the edge of the tissue. Logical operator with default value is T. If user does not want to perfom this action on their data then set to 'F'.}
}
\value{
two-dimensional matrix where total rows equal to total mass spectra in MSI data and total columns equal to bins created.
}
\description{
This function does peaks selection, removal of less frequent peaks and high intensity pixels from the tissue edge in MSI data. 

For the peaks selection, first the adaptive bins are created based on maximum intenstiy reference spectrum from single or multiple MSI datasets.
After bins creation, peaks selection performed from individual MSI dataset using  
local maxima search method. Threshold value to exclude the noisy peaks in the peaks selection step is automatically selected from ion-intensity region of the individual mass spectrum.
}
\examples{
folderpath <- 'MSIdata//AVA-PTX'
PreprocessedMSIData <- PreprocessingMSIData(folderpath) # Does all default tasks:peaks selection, remove peaks with less coverage area and does edge correction
PreprocessedMSIData <- PreprocessingMSIData(folderpath, removePeaks=0.0) # Does peaks selection and edge correction. 
PreprocessedMSIData <- PreprocessingMSIData(folderpath,removePeaks=0.0,edgecorrection=F ) # Only peaks selection 
IonInensityMatrix   <- PreprocessedMSIData[[1]][[1]]
ycord               <- PreprocessedMSIData[[1]][[3]]; 
xcord               <- PreprocessedMSIData[[1]][[2]]; 
SingleionImage      <- IonInensityMatrix[,41]
dim(SingleionImage) <- c(ycord,xcord)
image(SingleionImage,axes=FALSE,col=rainbow(256))

}
