% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatialvariablesselection.r
\name{VariablesSelection}
\alias{VariablesSelection}
\title{Variables selection}
\usage{
VariablesSelection(
  Inputdata,
  classids,
  method = "OLS",
  lagdist = 5,
  xycoord = 0
)
}
\arguments{
\item{Inputdata}{ion intensity matrix}

\item{classids}{class labels for variables selection}

\item{method}{method used for variables selection, 'OLS', 'SL', 'SE'.}

\item{lagdist}{the range of autocorrelation in spatial weight matrix for spatial methods}

\item{xcord, ycord}{the xy co-ordinates of the MSI ion image}

\item{maskimage}{exclude background pixels during variables selection}
}
\value{
regression coefficients and p-value for all variables
}
\description{
This function does variables selection using either spatial or non-spatial regression method. 
At the moment, function does variables selection for two class problem only where class of interest label as 2.
}
